gateway/requirements.txt
Flask==1.0.2
flask-nameko==1.4.0
flask-restful==0.3.7
nameko==2.11.0
python-dotenv==0.9.1
setuptools==41.0.0
apispec==0.38
flasgger==0.9.2